\name{fTrading-package}


\alias{fTrading-package}
\alias{fTrading}


\docType{package}


\title{Trading and Rebalancing Financial Instruments}


\description{

    The Rmetrics "fTrading" package is a collection of functions 
    for trading and rebalancing financial instruments.

}


\details{

    \tabular{ll}{
        Package:    \tab fTrading\cr
        Type:       \tab Package\cr
        Version:    \tab R 3.0.1\cr
        Date:       \tab 2014\cr
        License:    \tab GPL Version 2 or later\cr
        Copyright:  \tab (c) 1999-2014 Rmetrics Association\cr
        Repository: \tab R-FORGE\cr
        URL:        \tab \url{https://www.rmetrics.org}
    }

}


%\section{1 Introduction}{
%  The package includes three sections, benchmark analysis,
%  rolling analysis, and technical analysis.
%}


%\section{2 Benchmark Analysis}{
%  OBSOLETE ...
%}


%\section{3 Rolling Analysis}{
%  OBSOLETE ...
%}


\section{Content}{

    \emph{Utility Functions:}
    
    \preformatted{
    emaTA           Exponential Moving Average
    biasTA          Bias Indicator
    medpriceTA      Medium Price Indicator
    typicalpriceTA  Typical Price Indicator
    wcloseTA        Weighted Close Indicator
    rocTA           Rate of Change
    oscTA           Oscillator Indicator
    }
    
    
    \emph{Oscillator Indicators:}
    
    \preformatted{
    momTA           Momentum Indicator
    macdTA          MACD Indicator
    cdsTA           MACD Signal Line
    cdoTA           MACD Oscillator
    vohlTA          High/Low Volatility
    vorTA           Volatility Ratio
    }
    
    \preformatted{
    stochasticTA    Stochastics Oscillator
    fpkTA           Fast Percent K
    fpdTA           Fast Percent D
    spdTA           Slow Percent D
    apdTA           Averaged Percent D
    wprTA           William's Percent R
    rsiTA           Relative Strength Index
    }
    
    \emph{S-Plus Like Moving Averages:}
    
    \preformatted{
    SMA             Simple Moving Average
    EWMA            Exponentially Weighted  Moving Average
    }

}


\section{About Rmetrics:}{

    The \code{fTrading} Rmetrics package is written for educational 
    support in teaching "Computational Finance and Financial Engineering" 
    and licensed under the GPL.
    
}


\keyword{package}

